requirements.txt
pillow
opencv
numpy
matplotlib
keras
tensorflow